fix issue with not reported sim cards
There are devices for which ModemManager appears to not report single SIM cards
in the list of SIM slots. Sadly, there is no documentation on how SIM cards are
reported or not.

This extends the modem manager feature to first check for the SIM slots and to
check the active SIM as a fallback if no SIM slots are reported.