fix: modem manager support (#86)
We adapt for version 0.3.0 of the modemmanager crate. With this we
resolve issues with modemmanager replies that could not be parsed
properly. We resolve this by keeping the bearer information as it is and
directly converting it to JSON, without an additional conversion into an
internal data structure. This will hopefully make the modem info feature
more robust in the future.