fix: issue with high CPU load on SSH tunnel open (#137)
There was an issue with ssh causing a high CPU utilization upon creating
the SSH tunnel. The issue laid within the way ssh polls its file
descriptors: in the omnect-device-service we would wait for an initial
message that the ssh tunnel has been successfully created and then
implicitly drop and thus close the pipe to the ssh process. The ssh
process, on the other side, polls on its FDs for events, most notably in
this case stdout. The closed pipe would then cause this poll to always
directly fail with a POLLERR, thus causing ssh to poll at a high
frequency.

With this PR we take care to keep the pipe open until the ssh process
terminates.

Further changes (lint):
- broken format strings
- improved iterator access