fix(reboot-reason): log update validation failure only for erroneous termination (#160)
NOTE: this change also contains a FIXME in omnect-device-service's main
function which addresses that an explicit exit call at the end should
actually not be required.
To help further investigating on that issue an environment variable
named DONT_EXPLICITLY_EXIT_ON_TERMINATION can be set to skip the exit
call.

---------

Signed-off-by: Harry Waschkeit <44188360+HarryWaschkeit@users.noreply.github.com>